Update README.md
I. Motivation

    To give you experience in using arrays, pointers, structs, enums, and different I/O streams and writing program that takes arguments.

    To let you have fun with an application that is extremely captivating.

II. Introduction

The simple world program we will write for this project simulates a number of creatures running around in a simple square world. The world is an m-by-n two-dimensional grid of squares (The number m represents the height of the grid and the number n represents the width of the grid.). Each creature lives in one of the squares, faces in one of the major compass directions (north, east, south, or west) and belongs to a particular species, which determines how that creature behaves.

fl

la

ho

                        ho
                        	

                        fl

Figure 1. A 4-by-4 grid, which contains five creatures. Two creatures belong to the species flytrap (whose short name is “fl”), two belong to the species hop (whose short name is “ho”), and one belongs to the species landmine (whose short name is “la”). The direction of each creature is represented by the direction of the arrow.

represented by the direction of the arrow. For example, the flytrap at the top row is facing east and the flytrap at the bottom row is facing west.

Table 1. The list of instructions and their explanations.
			

The creature moves forward as long as the square it is facing is empty. If

hop
			

moving forward would put the creature outside the boundaries of the grid or
		

would cause it to land on top of another creature, the hop instruction does
		
			

nothing.
			

left
			

The creature turns left 90 degrees to face in a new direction.
			

right
			

The creature turns right 90 degrees to face in a new direction.
			
			

If the square immediately in front of this creature is occupied by a creature of
			

a different species (an “enemy”), that enemy creature is infected to become
			

the same as the infecting species. When a creature is infected, it keeps its

infect
			

position and orientation, but changes its internal species indicator and begins
		

executing the same program as the infecting creature, starting at step 1. If the
		
			

square immediately in front of this creature is empty, outside the grid, or
			

occupied by a creature of the same species, the infect instruction does
			

nothing.
			
			

If the square in front of the creature is inside the grid boundary and

ifempty
	

n
	

unoccupied, jump to step n of the program; otherwise, go on with the next
			

instruction in sequence.
			
			

If the creature is facing the border of the grid (which we imagine as

ifwall
	

n
		

consisting of a huge wall) jump to step n of the program; otherwise, go on
			

with the next instruction in sequence.
			

ifsame
	

n
		

If the square the creature is facing is occupied by a creature of the same
	

species, jump to step n; otherwise, go on with the next instruction.
		
			

ifenemy
	

n
	

If the square the creature is facing is occupied by a creature of an enemy

species, jump to step n; otherwise, go on with the next instruction.
		
			

go n
			

This instruction always jumps to step n, independent of any condition.